% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{build_database}
\alias{build_database}
\title{Construct a database tibble give a data path}
\usage{
build_database(path, pattern = "*.tif", ...)
}
\arguments{
\item{path}{character the directory to catalog}

\item{pattern}{character, the filename pattern (as glob) to search for}

\item{...}{other arguments for \code{\link{decompose_filename}}}
}
\value{
tibble database
}
\description{
Construct a database tibble give a data path
}
