% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{compose_filename}
\alias{compose_filename}
\title{Compose a filename from a database}
\usage{
compose_filename(x, path = ".", ext = ".tif")
}
\arguments{
\item{x}{database (tibble) with date, anomaly, version}

\item{path}{character, the root path for the filename}

\item{ext}{character, the filename extension to apply (with dot)}
}
\value{
character vector of filenames in form \code{<path>/ersst.vv.YYYYmm.ext}
}
\description{
Compose a filename from a database
}
