% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{decompose_filename}
\alias{decompose_filename}
\title{Decompose a filename into a database}
\usage{
decompose_filename(
  x = c("ersst.v5.185401.tif", "erssta.v5.185401.tif"),
  ext = ".tif"
)
}
\arguments{
\item{x}{character, vector of one or more filenames}

\item{ext}{character, the extension to remove (including dot)}
}
\value{
table (tibble) database of date, var, and version
}
\description{
Decompose a filename into a database
}
