% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdc.R
\name{download_ersst}
\alias{download_ersst}
\title{Download NCDC ERSST file to the specified directory.}
\usage{
download_ersst(
  uri = "https://www.ncei.noaa.gov/pub/data/cmb/ersst/v5/netcdf/ersst.v5.185401.nc",
  path = tempdir()
)
}
\arguments{
\item{uri}{character the URI of the file to downlaod with \code{download.file}}

\item{path}{charcater the name of the path to download to}
}
\value{
named logical where TRUE indicates success, name is the filename
}
\description{
Download NCDC ERSST file to the specified directory.
}
