% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\name{ersst_res}
\alias{ersst_res}
\title{Retrieve the spatial resolution stated in the global metadata}
\usage{
ersst_res(X)
}
\arguments{
\item{X}{ncdf4 object}
}
\value{
2 element [lon, lat] resolution vector
}
\description{
Retrieve the spatial resolution stated in the global metadata
}
