% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\name{ersst_t0}
\alias{ersst_t0}
\title{Retrieve the time epoch stated in the dimension attribute}
\usage{
ersst_t0(X)
}
\arguments{
\item{X}{ncdf4 object}
}
\value{
POSIXct time epoch
}
\description{
Retrieve the time epoch stated in the dimension attribute
}
