% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdf.R
\name{ersst_time}
\alias{ersst_time}
\title{Retrieve the time dimension of the NCDF object}
\usage{
ersst_time(X, t0 = ersst_t0(X), form = c("Date", "POSIXct")[2])
}
\arguments{
\item{X}{ncdf4 object}

\item{t0}{POSIXct, the origin of the time dimension}

\item{form}{character, output format. One of "Date" or "POSIXct" (default)}
}
\value{
vector fo time stamps as determined by \code{form} argument
}
\description{
Retrieve the time dimension of the NCDF object
}
