% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ersst.R
\name{ersst_url}
\alias{ersst_url}
\title{Craft a ERSST URL for a given date}
\usage{
ersst_url(
  date = Sys.Date() - 2,
  version = "v5",
  root = file.path("https://www.ncei.noaa.gov/pub/data/cmb/ersst")
)
}
\arguments{
\item{date}{character, POSIXt or Date the date to retrieve}

\item{version}{character the version to fetch}

\item{root}{character, the root URL}
}
\value{
one or more URLs
}
\description{
Craft a ERSST URL for a given date
}
