% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_geometry_type}
\alias{get_geometry_type}
\title{Get geometry type code}
\usage{
get_geometry_type(x, recursive = FALSE)
}
\arguments{
\item{x}{bbox, sf or sfc object}

\item{recursive}{logical, if TRUE drill down to get the type for each
feature.}
}
\value{
character vector such as "POINT" or "POLYGON"
}
\description{
Get geometry type code
}
