% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdc.R
\name{ncdc_list_uri}
\alias{ncdc_list_uri}
\title{Returns the known base URIs for file listings}
\usage{
ncdc_list_uri(version = "v5", append_slash = TRUE)
}
\arguments{
\item{version, }{character, ala "v5"}

\item{append_slash}{lofical, if TRUE append "/" to the end}
}
\value{
character uri
}
\description{
Returns the known base URIs for file listings
}
