% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdc.R
\name{ncdc_uri}
\alias{ncdc_uri}
\title{Returns the URI for people to browse}
\usage{
ncdc_uri(
  version = "v5",
  base_uri = "https://www.ncdc.noaa.gov/data-access/marineocean-data"
)
}
\arguments{
\item{version}{character, ala "v5"}

\item{base_uri}{character, the NCDC/NCEI base URI}
}
\value{
character uri
}
\description{
Returns the URI for people to browse
}
