% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{read_database}
\alias{read_database}
\title{Read a file-list database}
\usage{
read_database(path, filename = "database.csv.gz")
}
\arguments{
\item{path}{character the directory with the database}

\item{filename}{character, optional filename}
}
\value{
a tibble
}
\description{
Read a file-list database
}
