% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ersst.R
\name{read_ersst}
\alias{read_ersst}
\title{Read one or more local ERSST files}
\usage{
read_ersst(db, path)
}
\arguments{
\item{db}{tabular database}

\item{path}{chr, the data path}
}
\value{
stars object, with two attributes (variables) if both anomalies and raw values 
   are requested.  To read multiple attributes, the same number of dates
   must be available for each attribute otherwise an error is encountered.
}
\description{
Read one or more local ERSST files
}
