% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_sab}
\alias{read_sab}
\title{Read the South Atlantic Bight bout location data}
\usage{
read_sab(filename = system.file("extdata/sab.gpkg", package = "ersst"))
}
\arguments{
\item{filename}{charcater the name of the file}
}
\value{
sf POINT object
}
\description{
Read the South Atlantic Bight bout location data
}
