% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdc.R
\name{fetch_ersst}
\alias{fetch_ersst}
\title{Fetch a NCDC ERSST file, unpack it and save to the specified path}
\usage{
fetch_ersst(
  date = Sys.Date(),
  version = "v5",
  path = "./ersst_data",
  avail_db = NULL,
  overwrite = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{date}{Date or Date-castable, one of more dates to fetch}

\item{version}{character, the version}

\item{path}{character, the output path. Version subdirectory and
years subdirectories are made by default.}

\item{avail_db}{tibble, the current listing available ala \code{ncdc_list_available}
If not provided then \code{ncdc_list_available} is called.}

\item{overwrite}{logical, if TRUE overwrite existing file(s)}

\item{verbose}{logical, if TRUE then output messages}
}
\value{
a database tibble (date and version)
}
\description{
Fetch a NCDC ERSST file, unpack it and save to the specified path
}
