% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdc.R
\name{ncdc_list_available}
\alias{ncdc_list_available}
\title{Retrieve a listing of the \code{filenames} currently posted by NCDC/NCEI
of available ERSST.}
\usage{
ncdc_list_available(version = "v5", simplify = TRUE)
}
\arguments{
\item{version}{character, ala "v5"}

\item{simplify}{logical, if TRUE then make the returned tibble easier tp work with}
}
\value{
tibble of posted files (includes columns 'Name', 'Last modified', 'Size' and 'Description') unless \code{simply = TRUE} in which case
}
\description{
Retrieve a listing of the \code{filenames} currently posted by NCDC/NCEI
of available ERSST.
}
